pyproject.toml
[project]
name = "ffb-analytics"
version = "0.1.0"
description = "Fantasy Football cross-platform analytics dashboard"
requires-python = ">=3.10"
dependencies = [
    "espn-api>=0.38.0",
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dashboard = "src.dashboard:main"
